rename number to counter, add service layer

+++ testtask/src/main/java/com/pavelko/testtask/repository/NumberRepository.java
@@ -4,5 +4,5 @@ package com.pavelko.testtask.repository;
 import com.pavelko.testtask.model.NumberEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface NumberRepository extends JpaRepository<NumberEntity, Long> {
+public interface NumberRepository extends JpaRepository<NumberEntity, Integer> {
 }
